Change `Premade::install_all_handlers` to reset counters & flag.
As the "premade" offering, this is more useful, because most applications that do uninstalling
followed by re-installing will want the counters reset and will need the continue-flag to be
reset.

Any rare applications that for some reason want the counters to not be reset when re-installing
can still use the DIY mechanisms of this crate to setup something like that.