Clarify `README.md` regarding contrasting the "self-pipe trick".
The previous wording was accidentally probably giving a misimpression of the "self-pipe trick".  I
was trying to be too succinct with glossing over multiple different approaches to that trick
implicitly in my contrasting, and my wording might've been unfair to the most robust approach to
it (where the pipe's capacity can't cause issues when used only for just waking a thread).