Fix `util::abort::ewrite`.
`util::abort` was/is never called actually (because it's only used like `debug_assert!` failure or
`unreachable!`), and so the bug, fixed by this commit, never could occur for previous versions.
It's unnecessary to update which version of this crate you use just to have this fix.

(All of the other more-important source-code received much more of my careful attention to avoiding
bugs than `util::abort` had.)